+++ tasks/seq/makadrai_a_sobel/func_tests/main.cpp
@@ -12,10 +12,10 @@
 #include "seq/makadrai_a_sobel/include/ops_seq.hpp"
 
 TEST(makadrai_a_sobel_seq, test_2_2) {
-  size_t height_img = 2;
-  size_t width_img = 2;
-  std::vector<size_t> img = {100, 50, 150, 200};
-  std::vector<size_t> res(4);
+  int height_img = 2;
+  int width_img = 2;
+  std::vector<int> img = {100, 50, 150, 200};
+  std::vector<int> res(4);
 
   auto task_data = std::make_shared<ppc::core::TaskData>();
 
@@ -28,7 +28,7 @@ TEST(makadrai_a_sobel_seq, test_2_2) {
   task_data->outputs_count.emplace_back(width_img);
   task_data->outputs_count.emplace_back(height_img);
 
-  std::vector<size_t> ans = {228, 255, 201, 175};
+  std::vector<int> ans = {228, 255, 201, 175};
 
   makadrai_a_sobel_seq::Sobel sobel(task_data);
   ASSERT_TRUE(sobel.Validation());
@@ -39,10 +39,10 @@ TEST(makadrai_a_sobel_seq, test_2_2) {
 }
 
 TEST(makadrai_a_sobel_seq, test_1_1) {
-  size_t height_img = 1;
-  size_t width_img = 1;
-  std::vector<size_t> img = {100};
-  std::vector<size_t> res(1);
+  int height_img = 1;
+  int width_img = 1;
+  std::vector<int> img = {100};
+  std::vector<int> res(1);
 
   auto task_data = std::make_shared<ppc::core::TaskData>();
 
@@ -55,7 +55,7 @@ TEST(makadrai_a_sobel_seq, test_1_1) {
   task_data->outputs_count.emplace_back(width_img);
   task_data->outputs_count.emplace_back(height_img);
 
-  std::vector<size_t> ans = {0};
+  std::vector<int> ans = {0};
 
   makadrai_a_sobel_seq::Sobel sobel(task_data);
   ASSERT_TRUE(sobel.Validation());
@@ -66,10 +66,10 @@ TEST(makadrai_a_sobel_seq, test_1_1) {
 }
 
 TEST(makadrai_a_sobel_seq, test_3_3) {
-  size_t height_img = 3;
-  size_t width_img = 3;
-  std::vector<size_t> img = {32, 21, 61, 201, 231, 61, 132, 61, 111};
-  std::vector<size_t> res(9);
+  int height_img = 3;
+  int width_img = 3;
+  std::vector<int> img = {32, 21, 61, 201, 231, 61, 132, 61, 111};
+  std::vector<int> res(9);
 
   auto task_data = std::make_shared<ppc::core::TaskData>();
 
@@ -82,7 +82,7 @@ TEST(makadrai_a_sobel_seq, test_3_3) {
   task_data->outputs_count.emplace_back(width_img);
   task_data->outputs_count.emplace_back(height_img);
 
-  std::vector<size_t> ans = {235, 248, 152, 203, 121, 191, 247, 255, 170};
+  std::vector<int> ans = {235, 248, 152, 203, 121, 191, 247, 255, 170};
 
   makadrai_a_sobel_seq::Sobel sobel(task_data);
   ASSERT_TRUE(sobel.Validation());
@@ -93,10 +93,10 @@ TEST(makadrai_a_sobel_seq, test_3_3) {
 }
 
 TEST(makadrai_a_sobel_seq, test_error_1) {
-  size_t height_img = 3;
-  size_t width_img = 3;
-  std::vector<size_t> img = {32, 21, 61, 201, 231, 61, 132, 61, 111};
-  std::vector<size_t> res(9);
+  int height_img = 3;
+  int width_img = 3;
+  std::vector<int> img = {32, 21, 61, 201, 231, 61, 132, 61, 111};
+  std::vector<int> res(9);
 
   auto task_data = std::make_shared<ppc::core::TaskData>();
 
@@ -109,17 +109,17 @@ TEST(makadrai_a_sobel_seq, test_error_1) {
   task_data->outputs_count.emplace_back(width_img + 1);
   task_data->outputs_count.emplace_back(height_img);
 
-  std::vector<size_t> ans = {235, 248, 152, 203, 121, 191, 247, 255, 170};
+  std::vector<int> ans = {235, 248, 152, 203, 121, 191, 247, 255, 170};
 
   makadrai_a_sobel_seq::Sobel sobel(task_data);
   ASSERT_FALSE(sobel.Validation());
 }
 
 TEST(makadrai_a_sobel_seq, test_error_2) {
-  size_t height_img = 3;
-  size_t width_img = 3;
-  std::vector<size_t> img = {32, 21, 61, 201, 231, 61, 132, 61, 111};
-  std::vector<size_t> res(9);
+  int height_img = 3;
+  int width_img = 3;
+  std::vector<int> img = {32, 21, 61, 201, 231, 61, 132, 61, 111};
+  std::vector<int> res(9);
 
   auto task_data = std::make_shared<ppc::core::TaskData>();
 
@@ -132,17 +132,17 @@ TEST(makadrai_a_sobel_seq, test_error_2) {
   task_data->outputs_count.emplace_back(width_img);
   task_data->outputs_count.emplace_back(height_img);
 
-  std::vector<size_t> ans = {235, 248, 152, 203, 121, 191, 247, 255, 170};
+  std::vector<int> ans = {235, 248, 152, 203, 121, 191, 247, 255, 170};
 
   makadrai_a_sobel_seq::Sobel sobel(task_data);
   ASSERT_FALSE(sobel.Validation());
 }
 
 TEST(makadrai_a_sobel_seq, test_error_3) {
-  size_t height_img = 3;
-  size_t width_img = 3;
-  std::vector<size_t> img = {32, 21, 61, 201, 231, 61, 132, 61, 111};
-  std::vector<size_t> res(8);
+  int height_img = 3;
+  int width_img = 3;
+  std::vector<int> img = {32, 21, 61, 201, 231, 61, 132, 61, 111};
+  std::vector<int> res(8);
 
   auto task_data = std::make_shared<ppc::core::TaskData>();
 
@@ -155,7 +155,7 @@ TEST(makadrai_a_sobel_seq, test_error_3) {
   task_data->outputs_count.emplace_back(width_img);
   task_data->outputs_count.emplace_back(height_img);
 
-  std::vector<size_t> ans = {235, 248, 152, 203, 121, 191, 247, 255, 170};
+  std::vector<int> ans = {235, 248, 152, 203, 121, 191, 247, 255, 170};
 
   makadrai_a_sobel_seq::Sobel sobel(task_data);
   ASSERT_FALSE(sobel.Validation());
