+++ tasks/seq/makadrai_a_sobel/src/ops_seq.cpp
@@ -11,9 +11,9 @@ bool makadrai_a_sobel_seq::Sobel::PreProcessingImpl() {
   img.resize((width_img + peding) * (height_img + peding));
   simg.resize(width_img * height_img, 0);
 
-  const auto* in = reinterpret_cast<size_t*>(task_data->inputs[0]);
+  const auto* in = reinterpret_cast<int*>(task_data->inputs[0]);
 
-  for (size_t i = 0; i < height_img; i++) {
+  for (int i = 0; i < height_img; i++) {
     std::copy(in + (i * width_img), in + ((i + 1) * width_img), img.begin() + ((i + 1) * (width_img + peding) + 1));
   }
   return true;
@@ -27,8 +27,8 @@ bool makadrai_a_sobel_seq::Sobel::ValidationImpl() {
 
 bool makadrai_a_sobel_seq::Sobel::RunImpl() {
   int max_z = 1;
-  for (size_t i = 1; i < height_img + 1; i++) {
-    for (size_t j = 1; j < width_img + 1; j++) {
+  for (int i = 1; i < height_img + 1; i++) {
+    for (int j = 1; j < width_img + 1; j++) {
       int G_x = -1 * img[(i - 1) * (width_img + peding) + (j - 1)] +
                 -1 * img[(i - 1) * (width_img + peding) + (j + 1)] - 2 * img[(i - 1) * (width_img + peding) + j] +
                 2 * img[(i + 1) * (width_img + peding) + j] + 1 * img[(i + 1) * (width_img + peding) + (j - 1)] +
@@ -44,8 +44,8 @@ bool makadrai_a_sobel_seq::Sobel::RunImpl() {
     }
   }
 
-  for (size_t i = 0; i < width_img; i++) {
-    for (size_t j = 0; j < height_img; j++) {
+  for (int i = 0; i < width_img; i++) {
+    for (int j = 0; j < height_img; j++) {
       simg[i * height_img + j] = ((double)simg[i * height_img + j] / max_z) * 255;
     }
   }
@@ -54,6 +54,6 @@ bool makadrai_a_sobel_seq::Sobel::RunImpl() {
 }
 
 bool makadrai_a_sobel_seq::Sobel::PostProcessingImpl() {
-  std::copy(simg.begin(), simg.end(), reinterpret_cast<size_t*>(task_data->outputs[0]));
+  std::copy(simg.begin(), simg.end(), reinterpret_cast<int*>(task_data->outputs[0]));
   return true;
 }
